# Conflicts:
#	src/main/java/bitc/fullstack503/e2teamproject/controller/BoardController.java
#	src/main/java/bitc/fullstack503/e2teamproject/controller/HomeController.java
#	src/main/java/bitc/fullstack503/e2teamproject/repository/BoardRepository.java
#	src/main/java/bitc/fullstack503/e2teamproject/service/BoardService.java
#	src/main/java/bitc/fullstack503/e2teamproject/service/BoardServiceImpl.java

+++ src/main/java/bitc/fullstack503/e2teamproject/entity/UserEntity.java
@@ -34,7 +34,7 @@ public class UserEntity {
   @Column(name = "create_date")
   private LocalDateTime createDate = LocalDateTime.now();
 
-  private int level;
+  private byte level;
 
   @OneToMany(mappedBy = "user", cascade = CascadeType.ALL, fetch = FetchType.LAZY)
   @ToString.Exclude
